new app user with login page and register
