+++ it_zwierzak/homepage/urls.py
@@ -1,4 +1,5 @@
 from django.urls import path
+
 from homepage import views
 
 app_name = 'homepage'
@@ -8,6 +9,4 @@ urlpatterns = [
     path('statute/', views.statute, name='statute'),
     path('contact/', views.contact, name='contact'),
     path('advice/', views.advice, name='advice'),
-    path('login/', views.log_in, name='login'),
-    path('register/', views.register, name='register'),
 ]
