+++ it_zwierzak/it_zwierzak/settings.py
@@ -43,6 +43,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'homepage.apps.HomepageConfig',
+    'user.apps.UserConfig',
 ]
 
 MIDDLEWARE = [
