+++ it_zwierzak/it_zwierzak/urls.py
@@ -20,4 +20,5 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('adoption_announcement.urls')),
     path('', include('homepage.urls')),
+    path('', include('user.urls')),
     ]
