+++ it_zwierzak/user/forms.py
@@ -1,14 +1,8 @@
-from django.forms import ModelForm
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
-from django import forms
 
 
 class CreateUserForm(UserCreationForm):
     class Meta:
         model = User
         fields = {'username', 'email', 'password1', 'password2'}
-
-
-
-
