+++ it_zwierzak/user/templates/user/login.html
@@ -112,7 +112,7 @@
 
 				<div class="mt-4">
 					<div class="d-flex justify-content-center links">
-						Nie masz jeszcze profilu? <a href="{% url 'homepage:register' %}" class="ml-2">Zarejestruj</a>
+						Nie masz jeszcze profilu? <a href="{% url 'user:register' %}" class="ml-2">Zarejestruj</a>
 					</div>
 
 				</div>
